refactor(client): organize rbac, abac methods
