+++ client/src/abac/attribute/getGroupAttribute.ts
