+++ client/src/abac/index.ts
@@ -1,10 +1,10 @@
-export { checkAbacPolicy } from "./checkAbacPolicy";
-export { createPolicy } from "./createPolicy";
-export { deletePolicy } from "./deletePolicy";
-export { evaluatePolicies } from "./evaluatePolicies";
-export { getGroupAttribute } from "./getGroupAttribute";
-export { getPolicy } from "./getPolicy";
-export { getUserAttribute } from "./getUserAttribute";
-export { setGroupAttribute } from "./setGroupAttribute";
-export { setUserAttribute } from "./setUserAttribute";
-export { updatePolicy } from "./updatePolicy";
+export { checkAbacPolicy } from "./policy/checkAbacPolicy";
+export { createPolicy } from "./policy/createPolicy";
+export { deletePolicy } from "./policy/deletePolicy";
+export { evaluatePolicies } from "./policy/evaluatePolicies";
+export { getPolicy } from "./policy/getPolicy";
+export { updatePolicy } from "./policy/updatePolicy";
+export { getGroupAttribute } from "./attribute/getGroupAttribute";
+export { getUserAttribute } from "./attribute/getUserAttribute";
+export { setGroupAttribute } from "./attribute/setGroupAttribute";
+export { setUserAttribute } from "./attribute/setUserAttribute";
