+++ client/src/rbac/index.ts
@@ -1,10 +1,10 @@
-export * from "./addUserToGroup";
-export * from "./removeUserFromGroup";
-export * from "./assignPermissionToRole";
-export * from "./createPermission";
-export * from "./createRole";
-export * from "./getParentRole";
-export * from "./getRolePermissions";
-export * from "./setParentRole";
-export * from "./updateUserClaimsCache";
-export * from "./userHasPermission";
+export { addUserToGroup } from "./group/addUserToGroup";
+export { removeUserFromGroup } from "./group/removeUserFromGroup";
+export { assignPermissionToRole } from "./permission/assignPermissionToRole";
+export { createPermission } from "./permission/createPermission";
+export { userHasPermission } from "./permission/userHasPermission";
+export { createRole } from "./role/createRole";
+export { getParentRole } from "./role/getParentRole";
+export { getRolePermissions } from "./role/getRolePermissions";
+export { setParentRole } from "./role/setParentRole";
+export { updateUserClaimsCache } from "./cache/updateUserClaimsCache";
